removed router and fixed a bug was casuing issue
