+++ src/components/LoginForm.js
@@ -1,8 +1,6 @@
 import {
   Flex,
   Box,
-  Center,
-  Spacer,
   FormLabel,
   Input,
   Button,
